requirements.txt
requests
gradio
serpapi
feedparser
langchain-community
groq
python-dotenv
